Move uGMT params accessors to helper class
