+++ L1Trigger/L1TMuon/plugins/L1TMuonGlobalParamsESProducer.cc
@@ -29,6 +29,7 @@
 
 #include "CondFormats/L1TObjects/interface/L1TMuonGlobalParams.h"
 #include "CondFormats/DataRecord/interface/L1TMuonGlobalParamsRcd.h"
+#include "L1Trigger/L1TMuon/interface/L1TMuonGlobalParamsHelper.h"
 #include "L1Trigger/L1TMuon/interface/MicroGMTLUTFactories.h"
 
 //
@@ -64,26 +65,27 @@ L1TMuonGlobalParamsESProducer::L1TMuonGlobalParamsESProducer(const edm::Paramete
    // data is being produced
    setWhatProduced(this);
 
+   L1TMuonGlobalParamsHelper m_params_helper;
+
    // Firmware version
    unsigned fwVersion = iConfig.getParameter<unsigned>("fwVersion");
-
-   m_params.setFwVersion(fwVersion);
+   m_params_helper.setFwVersion(fwVersion);
 
    int bxMin = iConfig.getParameter<int>("bxMin");
    int bxMax = iConfig.getParameter<int>("bxMax");
    if (bxMin > bxMax) {
-      m_params.setBxMin(bxMax);
-      m_params.setBxMax(bxMin);
+      m_params_helper.setBxMin(bxMax);
+      m_params_helper.setBxMax(bxMin);
    } else {
-      m_params.setBxMin(bxMin);
-      m_params.setBxMax(bxMax);
+      m_params_helper.setBxMin(bxMin);
+      m_params_helper.setBxMax(bxMax);
    }
 
    bool disableCaloInputs = iConfig.getParameter<bool>("disableCaloInputs");
    if (disableCaloInputs) {
-      m_params.setCaloInputsToDisable(std::bitset<28>(0xFFFFFFF));
+      m_params_helper.setCaloInputsToDisable(std::bitset<28>(0xFFFFFFF));
    } else {
-      m_params.setCaloInputsToDisable(std::bitset<28>());
+      m_params_helper.setCaloInputsToDisable(std::bitset<28>());
    }
 
    std::vector<unsigned> bmtfInputsToDisable = iConfig.getParameter<std::vector<unsigned> >("bmtfInputsToDisable");
@@ -91,7 +93,7 @@ L1TMuonGlobalParamsESProducer::L1TMuonGlobalParamsESProducer(const edm::Paramete
    for (size_t i = 0; i < bmtfInputsToDisable.size(); ++i) {
      bmtfDisables.set(i, bmtfInputsToDisable[i] > 0);
    }
-   m_params.setBmtfInputsToDisable(bmtfDisables);
+   m_params_helper.setBmtfInputsToDisable(bmtfDisables);
 
    std::vector<unsigned> omtfInputsToDisable = iConfig.getParameter<std::vector<unsigned> >("omtfInputsToDisable");
    std::bitset<6> omtfpDisables;
@@ -103,8 +105,8 @@ L1TMuonGlobalParamsESProducer::L1TMuonGlobalParamsESProducer(const edm::Paramete
        omtfnDisables.set(i-6, omtfInputsToDisable[i] > 0);
      }
    }
-   m_params.setOmtfpInputsToDisable(omtfpDisables);
-   m_params.setOmtfnInputsToDisable(omtfnDisables);
+   m_params_helper.setOmtfpInputsToDisable(omtfpDisables);
+   m_params_helper.setOmtfnInputsToDisable(omtfnDisables);
 
    std::vector<unsigned> emtfInputsToDisable = iConfig.getParameter<std::vector<unsigned> >("emtfInputsToDisable");
    std::bitset<6> emtfpDisables;
@@ -116,22 +118,22 @@ L1TMuonGlobalParamsESProducer::L1TMuonGlobalParamsESProducer(const edm::Paramete
        emtfnDisables.set(i-6, emtfInputsToDisable[i] > 0);
      }
    }
-   m_params.setEmtfpInputsToDisable(emtfpDisables);
-   m_params.setEmtfnInputsToDisable(emtfnDisables);
+   m_params_helper.setEmtfpInputsToDisable(emtfpDisables);
+   m_params_helper.setEmtfnInputsToDisable(emtfnDisables);
 
 
-   m_params.setFwdPosSingleMatchQualLUTMaxDR(iConfig.getParameter<double>("FwdPosSingleMatchQualLUTMaxDR"));
-   m_params.setFwdNegSingleMatchQualLUTMaxDR(iConfig.getParameter<double>("FwdNegSingleMatchQualLUTMaxDR"));
-   m_params.setOvlPosSingleMatchQualLUTMaxDR(iConfig.getParameter<double>("OvlPosSingleMatchQualLUTMaxDR"));
-   m_params.setOvlNegSingleMatchQualLUTMaxDR(iConfig.getParameter<double>("OvlNegSingleMatchQualLUTMaxDR"));
-   m_params.setBOPosMatchQualLUTMaxDR(iConfig.getParameter<double>("BOPosMatchQualLUTMaxDR"), iConfig.getParameter<double>("BOPosMatchQualLUTMaxDREtaFine"));
-   m_params.setBONegMatchQualLUTMaxDR(iConfig.getParameter<double>("BONegMatchQualLUTMaxDR"), iConfig.getParameter<double>("BONegMatchQualLUTMaxDREtaFine"));
-   m_params.setFOPosMatchQualLUTMaxDR(iConfig.getParameter<double>("FOPosMatchQualLUTMaxDR"));
-   m_params.setFONegMatchQualLUTMaxDR(iConfig.getParameter<double>("FONegMatchQualLUTMaxDR"));
+   m_params_helper.setFwdPosSingleMatchQualLUTMaxDR(iConfig.getParameter<double>("FwdPosSingleMatchQualLUTMaxDR"));
+   m_params_helper.setFwdNegSingleMatchQualLUTMaxDR(iConfig.getParameter<double>("FwdNegSingleMatchQualLUTMaxDR"));
+   m_params_helper.setOvlPosSingleMatchQualLUTMaxDR(iConfig.getParameter<double>("OvlPosSingleMatchQualLUTMaxDR"));
+   m_params_helper.setOvlNegSingleMatchQualLUTMaxDR(iConfig.getParameter<double>("OvlNegSingleMatchQualLUTMaxDR"));
+   m_params_helper.setBOPosMatchQualLUTMaxDR(iConfig.getParameter<double>("BOPosMatchQualLUTMaxDR"), iConfig.getParameter<double>("BOPosMatchQualLUTMaxDREtaFine"));
+   m_params_helper.setBONegMatchQualLUTMaxDR(iConfig.getParameter<double>("BONegMatchQualLUTMaxDR"), iConfig.getParameter<double>("BONegMatchQualLUTMaxDREtaFine"));
+   m_params_helper.setFOPosMatchQualLUTMaxDR(iConfig.getParameter<double>("FOPosMatchQualLUTMaxDR"));
+   m_params_helper.setFONegMatchQualLUTMaxDR(iConfig.getParameter<double>("FONegMatchQualLUTMaxDR"));
 
    unsigned sortRankLUTPtFactor = iConfig.getParameter<unsigned>("SortRankLUTPtFactor");
    unsigned sortRankLUTQualFactor = iConfig.getParameter<unsigned>("SortRankLUTQualFactor");
-   m_params.setSortRankLUTFactors(sortRankLUTPtFactor, sortRankLUTQualFactor);
+   m_params_helper.setSortRankLUTFactors(sortRankLUTPtFactor, sortRankLUTQualFactor);
 
    auto absIsoCheckMemLUT = l1t::MicroGMTAbsoluteIsolationCheckLUTFactory::create (iConfig.getParameter<std::string>("AbsIsoCheckMemLUTPath"), fwVersion);
    auto relIsoCheckMemLUT = l1t::MicroGMTRelativeIsolationCheckLUTFactory::create (iConfig.getParameter<std::string>("RelIsoCheckMemLUTPath"), fwVersion);
@@ -152,46 +154,48 @@ L1TMuonGlobalParamsESProducer::L1TMuonGlobalParamsESProducer(const edm::Paramete
    auto oEtaExtrapolationLUT = l1t::MicroGMTExtrapolationLUTFactory::create (iConfig.getParameter<std::string>("OEtaExtrapolationLUTPath"), l1t::MicroGMTConfiguration::ETA_OUT, fwVersion);
    auto fEtaExtrapolationLUT = l1t::MicroGMTExtrapolationLUTFactory::create (iConfig.getParameter<std::string>("FEtaExtrapolationLUTPath"), l1t::MicroGMTConfiguration::ETA_OUT, fwVersion);
    auto rankPtQualityLUT = l1t::MicroGMTRankPtQualLUTFactory::create (iConfig.getParameter<std::string>("SortRankLUTPath"), fwVersion, sortRankLUTPtFactor, sortRankLUTQualFactor);
-   m_params.setAbsIsoCheckMemLUT(*absIsoCheckMemLUT);
-   m_params.setRelIsoCheckMemLUT(*relIsoCheckMemLUT);
-   m_params.setIdxSelMemPhiLUT(*idxSelMemPhiLUT);
-   m_params.setIdxSelMemEtaLUT(*idxSelMemEtaLUT);
-   m_params.setFwdPosSingleMatchQualLUT(*fwdPosSingleMatchQualLUT);
-   m_params.setFwdNegSingleMatchQualLUT(*fwdNegSingleMatchQualLUT);
-   m_params.setOvlPosSingleMatchQualLUT(*ovlPosSingleMatchQualLUT);
-   m_params.setOvlNegSingleMatchQualLUT(*ovlNegSingleMatchQualLUT);
-   m_params.setBOPosMatchQualLUT(*bOPosMatchQualLUT);
-   m_params.setBONegMatchQualLUT(*bONegMatchQualLUT);
-   m_params.setFOPosMatchQualLUT(*fOPosMatchQualLUT);
-   m_params.setFONegMatchQualLUT(*fONegMatchQualLUT);
-   m_params.setBPhiExtrapolationLUT(*bPhiExtrapolationLUT);
-   m_params.setOPhiExtrapolationLUT(*oPhiExtrapolationLUT);
-   m_params.setFPhiExtrapolationLUT(*fPhiExtrapolationLUT);
-   m_params.setBEtaExtrapolationLUT(*bEtaExtrapolationLUT);
-   m_params.setOEtaExtrapolationLUT(*oEtaExtrapolationLUT);
-   m_params.setFEtaExtrapolationLUT(*fEtaExtrapolationLUT);
-   m_params.setSortRankLUT(*rankPtQualityLUT);
+   m_params_helper.setAbsIsoCheckMemLUT(*absIsoCheckMemLUT);
+   m_params_helper.setRelIsoCheckMemLUT(*relIsoCheckMemLUT);
+   m_params_helper.setIdxSelMemPhiLUT(*idxSelMemPhiLUT);
+   m_params_helper.setIdxSelMemEtaLUT(*idxSelMemEtaLUT);
+   m_params_helper.setFwdPosSingleMatchQualLUT(*fwdPosSingleMatchQualLUT);
+   m_params_helper.setFwdNegSingleMatchQualLUT(*fwdNegSingleMatchQualLUT);
+   m_params_helper.setOvlPosSingleMatchQualLUT(*ovlPosSingleMatchQualLUT);
+   m_params_helper.setOvlNegSingleMatchQualLUT(*ovlNegSingleMatchQualLUT);
+   m_params_helper.setBOPosMatchQualLUT(*bOPosMatchQualLUT);
+   m_params_helper.setBONegMatchQualLUT(*bONegMatchQualLUT);
+   m_params_helper.setFOPosMatchQualLUT(*fOPosMatchQualLUT);
+   m_params_helper.setFONegMatchQualLUT(*fONegMatchQualLUT);
+   m_params_helper.setBPhiExtrapolationLUT(*bPhiExtrapolationLUT);
+   m_params_helper.setOPhiExtrapolationLUT(*oPhiExtrapolationLUT);
+   m_params_helper.setFPhiExtrapolationLUT(*fPhiExtrapolationLUT);
+   m_params_helper.setBEtaExtrapolationLUT(*bEtaExtrapolationLUT);
+   m_params_helper.setOEtaExtrapolationLUT(*oEtaExtrapolationLUT);
+   m_params_helper.setFEtaExtrapolationLUT(*fEtaExtrapolationLUT);
+   m_params_helper.setSortRankLUT(*rankPtQualityLUT);
 
    // LUT paths
-   m_params.setAbsIsoCheckMemLUTPath        (iConfig.getParameter<std::string>("AbsIsoCheckMemLUTPath"));
-   m_params.setRelIsoCheckMemLUTPath        (iConfig.getParameter<std::string>("RelIsoCheckMemLUTPath"));
-   m_params.setIdxSelMemPhiLUTPath          (iConfig.getParameter<std::string>("IdxSelMemPhiLUTPath"));
-   m_params.setIdxSelMemEtaLUTPath          (iConfig.getParameter<std::string>("IdxSelMemEtaLUTPath"));
-   m_params.setFwdPosSingleMatchQualLUTPath (iConfig.getParameter<std::string>("FwdPosSingleMatchQualLUTPath"));
-   m_params.setFwdNegSingleMatchQualLUTPath (iConfig.getParameter<std::string>("FwdNegSingleMatchQualLUTPath"));
-   m_params.setOvlPosSingleMatchQualLUTPath (iConfig.getParameter<std::string>("OvlPosSingleMatchQualLUTPath"));
-   m_params.setOvlNegSingleMatchQualLUTPath (iConfig.getParameter<std::string>("OvlNegSingleMatchQualLUTPath"));
-   m_params.setBOPosMatchQualLUTPath        (iConfig.getParameter<std::string>("BOPosMatchQualLUTPath"));
-   m_params.setBONegMatchQualLUTPath        (iConfig.getParameter<std::string>("BONegMatchQualLUTPath"));
-   m_params.setFOPosMatchQualLUTPath        (iConfig.getParameter<std::string>("FOPosMatchQualLUTPath"));
-   m_params.setFONegMatchQualLUTPath        (iConfig.getParameter<std::string>("FONegMatchQualLUTPath"));
-   m_params.setBPhiExtrapolationLUTPath     (iConfig.getParameter<std::string>("BPhiExtrapolationLUTPath"));
-   m_params.setOPhiExtrapolationLUTPath     (iConfig.getParameter<std::string>("OPhiExtrapolationLUTPath"));
-   m_params.setFPhiExtrapolationLUTPath     (iConfig.getParameter<std::string>("FPhiExtrapolationLUTPath"));
-   m_params.setBEtaExtrapolationLUTPath     (iConfig.getParameter<std::string>("BEtaExtrapolationLUTPath"));
-   m_params.setOEtaExtrapolationLUTPath     (iConfig.getParameter<std::string>("OEtaExtrapolationLUTPath"));
-   m_params.setFEtaExtrapolationLUTPath     (iConfig.getParameter<std::string>("FEtaExtrapolationLUTPath"));
-   m_params.setSortRankLUTPath              (iConfig.getParameter<std::string>("SortRankLUTPath"));
+   m_params_helper.setAbsIsoCheckMemLUTPath        (iConfig.getParameter<std::string>("AbsIsoCheckMemLUTPath"));
+   m_params_helper.setRelIsoCheckMemLUTPath        (iConfig.getParameter<std::string>("RelIsoCheckMemLUTPath"));
+   m_params_helper.setIdxSelMemPhiLUTPath          (iConfig.getParameter<std::string>("IdxSelMemPhiLUTPath"));
+   m_params_helper.setIdxSelMemEtaLUTPath          (iConfig.getParameter<std::string>("IdxSelMemEtaLUTPath"));
+   m_params_helper.setFwdPosSingleMatchQualLUTPath (iConfig.getParameter<std::string>("FwdPosSingleMatchQualLUTPath"));
+   m_params_helper.setFwdNegSingleMatchQualLUTPath (iConfig.getParameter<std::string>("FwdNegSingleMatchQualLUTPath"));
+   m_params_helper.setOvlPosSingleMatchQualLUTPath (iConfig.getParameter<std::string>("OvlPosSingleMatchQualLUTPath"));
+   m_params_helper.setOvlNegSingleMatchQualLUTPath (iConfig.getParameter<std::string>("OvlNegSingleMatchQualLUTPath"));
+   m_params_helper.setBOPosMatchQualLUTPath        (iConfig.getParameter<std::string>("BOPosMatchQualLUTPath"));
+   m_params_helper.setBONegMatchQualLUTPath        (iConfig.getParameter<std::string>("BONegMatchQualLUTPath"));
+   m_params_helper.setFOPosMatchQualLUTPath        (iConfig.getParameter<std::string>("FOPosMatchQualLUTPath"));
+   m_params_helper.setFONegMatchQualLUTPath        (iConfig.getParameter<std::string>("FONegMatchQualLUTPath"));
+   m_params_helper.setBPhiExtrapolationLUTPath     (iConfig.getParameter<std::string>("BPhiExtrapolationLUTPath"));
+   m_params_helper.setOPhiExtrapolationLUTPath     (iConfig.getParameter<std::string>("OPhiExtrapolationLUTPath"));
+   m_params_helper.setFPhiExtrapolationLUTPath     (iConfig.getParameter<std::string>("FPhiExtrapolationLUTPath"));
+   m_params_helper.setBEtaExtrapolationLUTPath     (iConfig.getParameter<std::string>("BEtaExtrapolationLUTPath"));
+   m_params_helper.setOEtaExtrapolationLUTPath     (iConfig.getParameter<std::string>("OEtaExtrapolationLUTPath"));
+   m_params_helper.setFEtaExtrapolationLUTPath     (iConfig.getParameter<std::string>("FEtaExtrapolationLUTPath"));
+   m_params_helper.setSortRankLUTPath              (iConfig.getParameter<std::string>("SortRankLUTPath"));
+
+   m_params = (L1TMuonGlobalParams)m_params_helper;
 }
